feat: show input under keyboard
