+++ src/screens/CreateTest/CreateTest.tsx
@@ -1,7 +1,7 @@
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {RootCreateTestParamsList} from 'types/navigation-types';
 import {TestSettings} from './TestSettings';
-import {QuestionsSettings} from './QuestionsSettings';
+import {QuestionsSettings} from './QuestionsSettings/index';
 import {ScreenList} from '../../navigation/navigation';
 import {ViewFlex} from '../../components/ui/ReadyStyles/Containers';
