+++ src/screens/CreateTest/QuestionsSettings/styles.ts
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  ViewContainer: {
+    flex: 1,
+    paddingTop: 30,
+    paddingHorizontal: 21,
+    paddingBottom: 3,
+  },
+  inner: {
+    flex: 1,
+    justifyContent: 'space-around',
+  },
+});
